nvmecontrol can fetch aws device name and volume id
A "device name" can be provided when AWS volumes are created.  On
newer instance that use NVMe disks, these names are ignored; names
like `/dev/nvmeX` are used instead and/but the actual device index
used is undefined (the order in which the volumes are attached is
undefined *and* may change between reboots).

The standard approach ([details here][ebsnvme-docs]) to bring
determinism to this process leverages the fact that AWS stores the
requested device name in the NVMe device's vendor extension space.
Amazon Linux provides a tool, ebsnvme-id, that fetches this
information.  It is also available in a raw form via the Linux `nvme
id-ctrl` command on sufficiently modern Linux distributions.

This commit extends the nvmecontrol application, adding an `aws`
sub-command that mimics the behavior of the `ebsnvme-id` tool.  E.g.

```
[root@freebsd /usr/src/sbin/nvmecontrol]# nvmecontrol aws nvme0
sda1
Volume ID: vol-0e711edd317b8a600
[root@freebsd /usr/src/sbin/nvmecontrol]#
```

[ebsnvme-docs]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html